# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Diploma RAG API service documentation.",
        "title": "Diploma API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/larek-tech/diploma/blob/api/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/domain/permissions/roles/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns list of roles permitted to domain.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain"
                ],
                "summary": "Get permitted roles.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Requested domain ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permitted roles",
                        "schema": {
                            "$ref": "#/definitions/pb.Domain"
                        }
                    },
                    "404": {
                        "description": "Domain not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates list of roles permitted to domain.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain"
                ],
                "summary": "Update permitted roles.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Requested domain ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New list of permitted roles",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.PermittedRoles"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated roles permissions",
                        "schema": {
                            "$ref": "#/definitions/pb.PermittedRoles"
                        }
                    },
                    "404": {
                        "description": "Domain not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/domain/permissions/users/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns list of users permitted to domain.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain"
                ],
                "summary": "Get permitted users.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Requested domain ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permitted users",
                        "schema": {
                            "$ref": "#/definitions/pb.Domain"
                        }
                    },
                    "404": {
                        "description": "Domain not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates list of users permitted to domain.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain"
                ],
                "summary": "Update permitted users.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Requested domain ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New list of permitted users",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.PermittedUsers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated users permissions",
                        "schema": {
                            "$ref": "#/definitions/pb.PermittedUsers"
                        }
                    },
                    "404": {
                        "description": "Domain not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/scenario/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns information about scenario.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scenario"
                ],
                "summary": "Get scenario.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Requested scenario ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Scenario",
                        "schema": {
                            "$ref": "#/definitions/pb.Scenario"
                        }
                    },
                    "400": {
                        "description": "Failed to get scenario",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Scenario not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update scenario information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scenario"
                ],
                "summary": "Update scenario.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Scenario ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update params",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateScenarioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Scenario updated",
                        "schema": {
                            "$ref": "#/definitions/pb.Scenario"
                        }
                    },
                    "400": {
                        "description": "Failed to update scenario",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Scenario not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete scenario by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scenario"
                ],
                "summary": "Delete scenario.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Scenario ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Scenario deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Failed to delete scenario",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Scenario not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.Domain": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "id": {
                    "type": "integer"
                },
                "scenarioIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "sourceIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                }
            }
        },
...
        "pb.LlmModel": {
            "type": "object",
            "properties": {
                "modelName": {
                    "type": "string"
                },
                "systemPrompt": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                },
                "topK": {
                    "type": "integer"
                },
                "topP": {
                    "type": "number"
                }
            }
        },
...
        "pb.MultiQuery": {
            "type": "object",
            "properties": {
                "nQueries": {
                    "description": "Количество перефразированных вопросов",
                    "type": "integer"
                },
                "queryModelName": {
                    "type": "string"
                },
                "useMultiquery": {
                    "type": "boolean"
                }
            }
        },
        "pb.PermittedRoles": {
            "type": "object",
            "properties": {
                "resourceId": {
                    "type": "integer"
                },
                "roleIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "pb.PermittedUsers": {
            "type": "object",
            "properties": {
                "resourceId": {
                    "type": "integer"
                },
                "userIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "pb.Reranker": {
            "type": "object",
            "properties": {
                "rerankerMaxLength": {
                    "type": "integer"
                },
                "rerankerModel": {
                    "type": "string"
                },
                "topK": {
                    "description": "Количество чанков после реранкинга",
                    "type": "integer"
                },
                "useRerank": {
                    "type": "boolean"
                }
            }
        },
...
        "pb.Scenario": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "domainId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "model": {
                    "$ref": "#/definitions/pb.LlmModel"
                },
                "multiQuery": {
                    "$ref": "#/definitions/pb.MultiQuery"
                },
                "reranker": {
                    "$ref": "#/definitions/pb.Reranker"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "vectorSearch": {
                    "$ref": "#/definitions/pb.VectorSearch"
                }
            }
        },
...
        "pb.UpdateScenarioRequest": {
            "type": "object",
            "properties": {
                "modelName": {
                    "type": "string"
                },
                "modelTopK": {
                    "type": "integer"
                },
                "nQueries": {
                    "description": "Количество перефразированных вопросов",
                    "type": "integer"
                },
                "queryModelName": {
                    "description": "Пока не знаю нучно ли будет",
                    "type": "string"
                },
                "rerankerMaxLength": {
                    "type": "integer"
                },
                "rerankerModel": {
                    "type": "string"
                },
                "rerankerTopK": {
                    "description": "Количество чанков после реранкинга",
                    "type": "integer"
                },
                "scenarioId": {
                    "type": "integer"
                },
                "searchByQuery": {
                    "type": "boolean"
                },
                "systemPrompt": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                },
                "threshold": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "topN": {
                    "description": "Сколько чанков забирать при векторном поиске.",
                    "type": "integer"
                },
                "topP": {
                    "type": "number"
                },
                "useMultiquery": {
                    "type": "boolean"
                },
                "useRerank": {
                    "type": "boolean"
                }
            }
        },
...
        "pb.VectorSearch": {
            "type": "object",
            "properties": {
                "searchByQuery": {
                    "type": "boolean"
                },
                "threshold": {
                    "type": "number"
                },
                "topN": {
                    "description": "Сколько чанков забирать при векторном поиске.",
                    "type": "integer"
                }
            }
        },
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. Negative\nsecond values with fractions must still have non-negative nanos values\nthat count forward in time. Must be from 0 to 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch\n1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to\n9999-12-31T23:59:59Z inclusive.",
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}